No errors, Print quotient

+++ print_quotient_two_numbers.py
@@ -1,5 +1,5 @@
 #Ask for 2 numbers
-num1 = int(input("Enter a number: "))
-num2 = int(input("Enter a number: "))
+num1 = float(input("Enter a number: "))
+num2 = float(input("Enter a number: "))
 #Print the quotient
 print("The quotient of the 2 numbers is: ", (num1 / num2))
